refactor: 주문 서비스 테스트 수정
